backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
aiomysql
python-jose[cryptography]
passlib
pydantic[email]
python-dotenv
